App Linking and Route Filter Bugfix
Routes could be shared by code previously. I added app linking so users can skip the step of having to manually enter the code.

Previously, filtering was removing the routes not just from the view, but also from the provider itself. This has been fixed by adding an invalidate call on the provider after a filter option is selected or deselected. This is a stopgap measure and the real issue may lie elsewhere.

Share button added to About App Screen which shares the app's App Store page link.

Added isHardcoded flag to hardcoded A/C bus routes.

Regrouped buttons on List Routes screen to distinguish between actions for all routes and actions for selected routes only.

+++ lib/domain/use_cases/routes_use_case.dart
@@ -26,6 +26,9 @@ class RoutesUseCase extends _$RoutesUseCase {
   }
 
   Future<void> deleteRoute(String routeName) async {
+    final hc = repo.getHardcoded();
+    final isRouteHc = hc.where((hcr) => hcr.name == routeName).isNotEmpty;
+    if (isRouteHc) throw '$routeName cannot be deleted. It is hardcoded';
     await repo.deleteRoute(routeName);
     state = AsyncData(
       state.value!
